+++ ios/Event+JSON.swift
@@ -340,6 +340,17 @@ extension VideoDownloadQualityChangedEvent {
     }
 }
 
+extension VideoPlaybackQualityChangedEvent {
+    func toJSON() -> [AnyHashable: Any] {
+        [
+            "newVideoQuality": RCTConvert.toJson(videoQuality: videoQualityNew),
+            "oldVideoQuality": RCTConvert.toJson(videoQuality: videoQualityOld),
+            "name": name,
+            "timestamp": timestamp
+        ]
+    }
+}
+
 extension CastStartedEvent {
     func toJSON() -> [AnyHashable: Any] {
         [
@@ -378,5 +389,16 @@ extension DownloadFinishedEvent {
             json["lastRedirectLocation"] = lastRedirectLocation.absoluteString
         }
         return json
+    }
+}
+
+extension PlaybackSpeedChangedEvent {
+    func toJSON() -> [AnyHashable: Any] {
+        [
+            "name": name,
+            "timestamp": timestamp,
+            "from": from,
+            "to": to,
+        ]
     }
 }