+++ ios/RNPlayerView+PlayerListener.swift
@@ -77,6 +77,10 @@ extension RNPlayerView: PlayerListener {
         onTimeChanged?(event.toJSON())
     }
 
+    public func onPlaybackSpeedChanged(_ event: PlaybackSpeedChangedEvent, player: Player) {
+        onPlaybackSpeedChanged?(event.toJSON())
+    }
+
     public func onSourceLoad(_ event: SourceLoadEvent, player: Player) {
         onSourceLoad?(event.toJSON())
     }
@@ -169,7 +173,11 @@ extension RNPlayerView: PlayerListener {
         onAdStarted?(event.toJSON())
     }
 
-    public func onVideoPlaybackQualityChanged(_ event: VideoDownloadQualityChangedEvent, player: Player) {
+    public func onVideoDownloadQualityChanged(_ event: VideoDownloadQualityChangedEvent, player: Player) {
+        onVideoDownloadQualityChanged?(event.toJSON())
+     }
+
+    public func onVideoPlaybackQualityChanged(_ event: VideoPlaybackQualityChangedEvent, player: Player) {
         onVideoPlaybackQualityChanged?(event.toJSON())
     }