+++ CHANGELOG.md
@@ -7,7 +7,14 @@
 - `LiveConfig.minTimeshiftBufferDepth` to control the minimum buffer depth of a stream needed to enable time shifting.
 - `Player.buffer` to control buffer preferences and to query the current buffer state
 - `DownloadFinishedEvent` to signal when the download of specific content has finished
+- `Player.videoQuality`, `Player.availableVideoQualities`, and `VideoDownloadQualityChangedEvent` to query current video qualities and listen to related changes
+- `Player.playbackSpeed`, `Player.canPlayAtPlaybackSpeed`, `PlaybackSpeedChangedEvent` to query, control, and listen to changes to the speed of the playback
 - Support for `UserInterfaceType.Subtitle` on Android
+
+### Fixed
+
+- Android: Playback doesn't pause when app goes to background
+- Android: `PlayerView.onDestroy` not being called when the view is detached from the view hierarchy
 
 ## [0.13.0] (2023-10-20)